revert extraction of chat completion manager
